Fix setting SIAddr of DHCP header:
As far as I can tell, rep.SetSIAddr(nextServer) was never actually
copying the nextServer (type net.IP) to the dhcp4.Packet.RawPacket.
The default length of net.IP is 16, put the SIAddr of the RawPacket
is only of length 4.

Up to this point, this hasn't been an issue because the TFTP server
has always been the same as the DHCP server. So the client wouldn't
get a SIAddr, so it would just use the DHCP server's address. And again,
since the TFTP and DHCP were both being served from the same IP it
was not a problem.

With the introduction of being able to specify a remote TFTP server
this came to light in testing. The net.IP.To4() method solves the issue,
both for remote and local TFTP serving.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>